Add the option for specifying the session max age through configuration
