+++ session.go
@@ -1,21 +1,14 @@
 package idp
 
 import (
-	"encoding/base64"
+	"errors"
 	"github.com/crewjam/saml"
 	"github.com/crewjam/saml/samlidp"
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
-	"html/template"
 	"net/http"
 	"time"
 )
-
-var sessionMaxAge = time.Hour * 24 * 14 // 14 days
-
-func init() {
-	saml.MaxIssueDelay = sessionMaxAge
-}
 
 func (s *Server) GetSession(w http.ResponseWriter, r *http.Request, req *saml.IdpAuthnRequest) *saml.Session {
 	if r.Method == http.MethodPost && r.PostForm.Get("username") != "" {
@@ -31,12 +24,13 @@ func (s *Server) GetSession(w http.ResponseWriter, r *http.Request, req *saml.Id
 		}
 
 		now := saml.TimeNow()
+		expires := now.Add(time.Duration(s.config.SessionMaxAge) * time.Minute)
 
 		session := &saml.Session{
 			ID:                    uuid.NewString(),
 			NameID:                user.Email,
 			CreateTime:            now,
-			ExpireTime:            now.Add(sessionMaxAge),
+			ExpireTime:            expires,
 			Index:                 uuid.NewString(),
 			UserName:              user.Name,
 			Groups:                user.Groups[:],
@@ -55,7 +49,7 @@ func (s *Server) GetSession(w http.ResponseWriter, r *http.Request, req *saml.Id
 		http.SetCookie(w, &http.Cookie{
 			Name:     "session",
 			Value:    session.ID,
-			MaxAge:   int(sessionMaxAge.Seconds()),
+			Expires:  expires,
 			HttpOnly: true,
 			Secure:   r.URL.Scheme == "https",
 			Path:     "/",
@@ -68,7 +62,7 @@ func (s *Server) GetSession(w http.ResponseWriter, r *http.Request, req *saml.Id
 		session, err := s.Store.GetSession(cookie.Value)
 
 		if err != nil {
-			if err == samlidp.ErrNotFound {
+			if errors.Is(err, samlidp.ErrNotFound) {
 				s.serveLoginPage(w, r, req, "")
 				return nil
 			}
@@ -87,48 +81,4 @@ func (s *Server) GetSession(w http.ResponseWriter, r *http.Request, req *saml.Id
 
 	s.serveLoginPage(w, r, req, "")
 	return nil
-}
-
-type LoginPageData struct {
-	Title       string
-	Description template.HTML
-	Users       []User
-	Toast       string
-	Username    string
-	Url         string
-	SamlRequest string
-	RelayState  string
-}
-
-func (s *Server) serveLoginPage(w http.ResponseWriter, r *http.Request, req *saml.IdpAuthnRequest, toast string) {
-	data := LoginPageData{
-		Title:       "Login",
-		Description: "",
-		Toast:       toast,
-		Username:    r.PostForm.Get("username"),
-		Url:         req.IDP.SSOURL.String(),
-		SamlRequest: base64.StdEncoding.EncodeToString(req.RequestBuffer),
-		RelayState:  req.RelayState,
-	}
-
-	options := s.config.LoginPageOptions
-
-	if options.Title != "" {
-		data.Title = options.Title
-	}
-
-	if options.Description != "" {
-		data.Description = renderMarkdown(options.Description)
-	}
-
-	if options.DumpUsers {
-		data.Users = s.config.Users
-	}
-
-	render := s.router.HTMLRender.Instance("login.html", data)
-
-	err := render.Render(w)
-	if err != nil {
-		panic(err)
-	}
 }
